backend/requirements.txt
# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# 认证安全
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# AI调用
openai>=1.0.0
dashscope>=1.14.0
aiohttp>=3.9.0

# PDF解析
pdfplumber>=0.10.0
pymupdf>=1.23.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 配置/工具
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
